Rename frames.rs to frame.rs to match equivalent C file
